pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-cleanup"
version = "0.1.0"
description = "Windows-first MP3 auto-tagger and organizer"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Trevor + Codex" }]
dependencies = [
  "pyacoustid>=1.3.0",
  "musicbrainzngs>=0.7.1",
  "mutagen>=1.47.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
music-cleanup = "music_cleanup.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
